+++ src/Core/ApiDoc/ApiDoc.php
@@ -3,8 +3,7 @@
 namespace FlyCubePHP\Core\ApiDoc;
 
 use Exception;
-use FlyCubePHP\Core\Config\Config as Config;
-use FlyCubePHP\Core\Error\ErrorAssetPipeline as ErrorAssetPipeline;
+use FlyCubePHP\Core\Config\Config;
 use FlyCubePHP\HelperClasses\CoreHelper;
 use FlyCubePHP\Core\Error\Error;
