+++ src/Core/AssetPipeline/CSSBuilder.php
@@ -11,8 +11,8 @@ namespace FlyCubePHP\Core\AssetPipeline;
 include_once 'CSSMinifier.php';
 include_once __DIR__.'/../../HelperClasses/CoreHelper.php';
 
-use \FlyCubePHP\Core\Config\Config as Config;
-use \FlyCubePHP\HelperClasses\CoreHelper as CoreHelper;
+use FlyCubePHP\Core\Config\Config;
+use FlyCubePHP\HelperClasses\CoreHelper;
 use FlyCubePHP\Core\Error\ErrorAssetPipeline;
 use ScssPhp\ScssPhp\OutputStyle;
