+++ src/Core/AssetPipeline/JSBuilder.php
@@ -10,10 +10,9 @@ namespace FlyCubePHP\Core\AssetPipeline;
 
 include_once __DIR__.'/../../HelperClasses/CoreHelper.php';
 
-//use Exception;
-use \FlyCubePHP\Core\Config\Config as Config;
-use \FlyCubePHP\HelperClasses\CoreHelper as CoreHelper;
-use \FlyCubePHP\Core\Error\ErrorAssetPipeline as ErrorAssetPipeline;
+use FlyCubePHP\Core\Config\Config;
+use FlyCubePHP\HelperClasses\CoreHelper;
+use FlyCubePHP\Core\Error\ErrorAssetPipeline;
 
 class JSBuilder
 {
