+++ src/Core/Logger/Logger.php
@@ -12,8 +12,8 @@ include_once __DIR__.'/../Error/Error.php';
 include_once __DIR__.'/../Routes/RouteCollector.php';
 include_once __DIR__.'/../../HelperClasses/CoreHelper.php';
 
-use \FlyCubePHP\Core\Config\Config as Config;
-use \FlyCubePHP\HelperClasses\CoreHelper as CoreHelper;
+use FlyCubePHP\Core\Config\Config;
+use FlyCubePHP\HelperClasses\CoreHelper;
 use Psr\Log\LoggerInterface;
 use Exception;
