+++ src/Core/Migration/MySQLMigrator.php
@@ -5,8 +5,7 @@ namespace FlyCubePHP\Core\Migration;
 include_once 'BaseMigrator.php';
 include_once __DIR__.'/../../HelperClasses/CoreHelper.php';
 
-use FlyCubePHP\Core\Database\BaseDatabaseAdapter;
-use \FlyCubePHP\HelperClasses\CoreHelper as CoreHelper;
+use FlyCubePHP\HelperClasses\CoreHelper;
 
 class MySQLMigrator extends BaseMigrator
 {
