+++ src/Core/Session/Session.php
@@ -14,8 +14,8 @@ namespace FlyCubePHP\Core\Session;
 include_once __DIR__.'/../Routes/RouteCollector.php';
 
 use Exception;
-use \FlyCubePHP\Core\Routes\RouteCollector as RouteCollector;
-use \FlyCubePHP\HelperClasses\CoreHelper;
+use FlyCubePHP\Core\Routes\RouteCollector;
+use FlyCubePHP\HelperClasses\CoreHelper;
 
 class Session
 {
